Reorgagnisation of all dirs + optimisations + readme.md + progress-bar
