+++ readme.md
@@ -1,0 +1,25 @@
+# Mac Roblox Utils
+
+#### Created by IAmEvan
+
+## Installation 🔽
+### Git Clone
+It will just clone the repo
+
+`git clone https://github.com/iamEvanYT/MacRobloxUtility.git`
+
+#### Install requirements
+
+`pip install tkinter psutil requests alive-progress` or `pip3...`
+
+### Pip
+`Not available`
+
+## Run ▶️ 🚀
+### Basic python 🐍
+
+`python main.py`
+
+### If doesn't work, try
+`py main.py` or
+`python3 main.py`
