+++ utils/constants.py
@@ -1,0 +1,12 @@
+BEHAVIOR_MAP = {
+    "F": "Fast",
+    "DF": "Dynamic Fast",
+    "SF": "Synchronized Fast"
+}
+
+TYPE_MAP = {
+    "Flag": "bool",
+    "Int": "int",
+    "String": "string",
+    "Log": "byte",
+}
